+++ src/UI/TopMenuContainer.gd
@@ -65,7 +65,8 @@ func setup_view_menu() -> void:
 		"Show Rulers" : InputMap.get_action_list("show_rulers")[0].get_scancode_with_modifiers(),
 		"Show Guides" : InputMap.get_action_list("show_guides")[0].get_scancode_with_modifiers(),
 		"Show Animation Timeline" : 0,
-		"Zen Mode" : InputMap.get_action_list("zen_mode")[0].get_scancode_with_modifiers()
+		"Zen Mode" : InputMap.get_action_list("zen_mode")[0].get_scancode_with_modifiers(),
+		"Fullscreen Mode" : InputMap.get_action_list("toggle_fullscreen")[0].get_scancode_with_modifiers(),
 		}
 	view_menu = Global.view_menu.get_popup()
 
@@ -234,8 +235,10 @@ func view_menu_id_pressed(id : int) -> void:
 			toggle_show_guides()
 		4: # Show animation timeline
 			toggle_show_anim_timeline()
-		5: # Fullscreen mode
+		5: # Zen mode
 			toggle_zen_mode()
+		6: # Fullscreen mode
+			toggle_fullscreen()
 
 	Global.canvas.update()
 
@@ -286,6 +289,11 @@ func toggle_zen_mode() -> void:
 	Global.control.get_node("MenuAndUI/UI/CanvasAndTimeline/ViewportAndRulers/TabsContainer").visible = zen_mode
 	zen_mode = !zen_mode
 	view_menu.set_item_checked(5, zen_mode)
+
+
+func toggle_fullscreen() -> void:
+	OS.window_fullscreen = !OS.window_fullscreen
+	view_menu.set_item_checked(6, OS.window_fullscreen)
 
 
 func image_menu_id_pressed(id : int) -> void: